moved register test forester to light-program-test
