+++ sdk-libs/program-test/src/utils/mod.rs
@@ -1,5 +1,6 @@
 pub mod assert;
 pub mod create_account;
 pub mod find_light_bin;
+pub mod register_test_forester;
 pub mod setup_light_programs;
 pub mod tree_accounts;
